+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 * `RedistributeSearch` heuristic
 
+### Changed
+
+* use experimental mode to print dynamic heuristic statistics
+
 
 ## [v1.18.1] - 2022-05-13